chat_api/requirements.txt
flask[async]>=2.0  
Werkzeug>=2.0
pandas
Flask-RESTful
Flask-SQLAlchemy
Flask-Marshmallow
Marshmallow
marshmallow-sqlalchemy
psycopg2-binary
flask-restplus
flask_swagger
flask_swagger_ui
pyyaml==6.0.1